Fixed: Videos not recording
Videos were not recording after the first motion detected video was recorded.

Another issue was when python mail package couldn't find the environment variables for the sender, the motion detected videos were stored and recorded but then after page refresh the sites live feed video would freeze cuz of the mail send error.

To fix this always add the environment variables so the email can be sent.
If it cant be sent and gives error it freezes the live video feed after page refresh.